requirements.txt
streamlit
pandas
numpy
scipy
tensorflow-cpu
pickle-mixin
scikit-learn == 1.0.2